requirements.txt
asgiref==3.5.2
Django==4.1.1
django-admin-sortable2==2.1.3
Pillow==9.2.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
